# api/docs/swagger.yaml
basePath: /api
definitions:
...
  docs.ErrorResponse:
    properties:
      error:
        example: Error message
        type: string
      success:
        example: false
        type: boolean
    type: object
...
  docs.SuccessResponse:
    properties:
      data: {}
      success:
        example: true
        type: boolean
    type: object
...
  handlers.CreateFileShareRequest:
    properties:
      isFolder:
        type: boolean
      itemName:
        type: string
      itemPath:
        type: string
      message:
        type: string
      permissionLevel:
        type: integer
      sharedWithId:
        type: string
    type: object
...
  handlers.FileShare:
    properties:
      createdAt:
        type: string
      id:
        type: integer
      isFolder:
        type: boolean
      itemName:
        type: string
      itemPath:
        type: string
      message:
        type: string
      ownerId:
        type: string
      ownerUsername:
        description: Additional fields for display
        type: string
      permissionLevel:
        type: integer
      sharedWithId:
        type: string
      sharedWithUsername:
        type: string
      updatedAt:
        type: string
    type: object
...
host: localhost:3080
info:
  contact:
    email: support@simplecloudvault.io
    name: API Support
  description: |-
    Enterprise-grade cloud file sharing and management API.
    SimpleCloudVault provides comprehensive file management, sharing, and collaboration features.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: SimpleCloudVault API
  version: "1.0"
paths:
...
  /file-shares:
    post:
      consumes:
      - application/json
      description: Share a file or folder with another user
      parameters:
      - description: File share request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateFileShareRequest'
      produces:
      - application/json
      responses:
        "200":
          description: File share created
          schema:
            allOf:
            - $ref: '#/definitions/docs.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/handlers.FileShare'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "409":
          description: Share already exists
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create file share
      tags:
      - FileShares
  /file/{path}:
    delete:
      consumes:
      - application/json
      description: Permanently delete a file by path
      parameters:
      - description: File path
        in: path
        name: path
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: File deleted successfully
          schema:
            $ref: '#/definitions/docs.SuccessResponse'
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: File not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete file
      tags:
      - Files
    put:
      consumes:
      - text/plain
      description: Save text content to an existing file (for text editor)
      parameters:
      - description: File path
        in: path
        name: path
        required: true
        type: string
      - description: File content
        in: body
        name: content
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: File saved successfully
          schema:
            $ref: '#/definitions/docs.SuccessResponse'
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: File not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Save file content
      tags:
      - Files
  /files/{path}:
    get:
      description: Download a file by path. Supports both inline viewing and forced
        download.
      parameters:
      - description: File path
        in: path
        name: path
        required: true
        type: string
      - description: Force download with Content-Disposition attachment
        in: query
        name: download
        type: boolean
      produces:
      - application/octet-stream
      responses:
        "200":
          description: File content
          schema:
            type: file
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: File not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Download file
      tags:
      - Download
...
securityDefinitions:
  BearerAuth:
    description: 'JWT Bearer token. Format: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"